[LC-556] move validate_jsonschema from validator to each dispatcher
